Fix bug: shutdown fails when ES not running
This fixes a bug that prevents safe shutdown when EmulationStation is not running.

The os.system() calls use && (logical and) statements, which require each command to return 0 on exit, or the next command won't be run. The "sudo killall emulationstation" fails when ES isn't running, such as when the user has pressed F4 to work at the terminal. Thus the system doesn't shut down. Using simple ";" statements sequences the commands without regard to exit status.