Add files via upload
This is a complete change for those doing more of a home brew approach and coding a single led and button to function for both a safe reboot and a safe shutdown.  I modified the source code from Retroflag so I wanted to give credit where it's due and create a fork and pull request so others can use the code for themselves if they like.  All the code is documented in the safeshutdown.py script.