Add scripts and instructions for blast16
I found the scripts didn't work correctly for [blast16](http://blast16.tripware.es/) (with the MEGAPi case), probably because the blast16 image is a little old at this point. It fails to connect to GitHub with  `ERROR: The certificate of 'raw.githubusercontent.com' is not yet activated.`

This PR adds a new blast16_install.sh script, a new blast16_SafeShutdown.py script to close blast16's UI nicely, and updates the readme with new instructions.